firmware code update
The LDAD is officially up and running!  Well the very first  release candidate of the firmware. Now, on this code revision,  electrical resistance and measurement voltage are correctly measured. And this means the smart DAQ is now ready to be installed on all specimens in a experimental campaign before configure with the correct calibration curves (temperature compensated). From now on all changes will be towards improving usability and user experience in a laboratory.